+++ .cz-config.js
@@ -11,10 +11,10 @@ module.exports = {
     { value: '🔧工具', name: '工具:    开发工具变动(构建、脚手架工具等)' },
     { value: '⏪回滚', name: '回滚:    代码回退' }],
   scopes: [
-    { name: '模块1' },
-    { name: '模块2' },
-    { name: '模块3' },
-    { name: '模块4' }
+    { name: 'UI' },
+    { name: 'Interaction' },
+    { name: 'Echarts' },
+    { name: 'Data' }
   ],
   // it needs to match the value for field type. Eg.: 'fix'
   /*  scopeOverrides: {