+++ src/layout/FloatBall/index.js
@@ -1,0 +1,3 @@
+import FloatBall from './index.vue';
+
+export { FloatBall };