+++ src/layout/Sidebar/index.vue
@@ -1,46 +1,74 @@
 <template>
-<n-layout-sider :width="220" bordered>
-
-</n-layout-sider>
-
+  <n-layout-sider
+    collapse-mode="transform"
+    :collapsed-width="0"
+    :width="280"
+    bordered
+    :collapsed="collapsed"
+    class="app-layout-sider"
+    show-trigger="bar"
+    style="border-top-right-radius: 16px"
+    @collapse="collapsed = true"
+    @expand="collapsed = false"
+  >
+    <span class="app-layout-sider__title"> How to use provider </span>
+    <n-menu
+      :value="activeName"
+      :options="layoutOptions"
+      :collapsed-width="64"
+      :collapsed-icon-size="22"
+      @update:value="handleMenuSelect"
+    />
+  </n-layout-sider>
 </template>
 
 <script setup>
+import { ref } from 'vue'
+const collapsed = ref(false)
+const activeName = ref('/')
+const layoutOptions = ref([
+	{
+	  label: 'Home',
+	  key: '/',
+	},
+	{
+	  label: 'Request',
+	  key: '/request',
+	},
+	{
+		label: 'Message',
+		key: '/message',
+	},
+	{
+		label: 'Dialog',
+		key: '/dialog',
+	},
+	{
+		label: 'Notification',
+		key: '/notification',
+	},
+	{
+		label: 'LoadingBar',
+		key: '/loadingBar',
+	},
+])
+const handleMenuSelect = (value) => {
+  activeName.value = value
   
-
+}
 </script>
 
 
 <style scoped>
-.logo {
-  position: sticky;
-  top: 0;
-  z-index: 1;
+.app-layout-sider__title {
   display: flex;
   align-items: center;
-  padding: 12px 20px;
-  /* border-bottom: 1px solid var(--border-color); */
-  background: var(--color);
-  font-size: 1.8em;
-  font-weight: 600;
-  line-height: 1;
-  text-decoration: none;
-  transition: padding 0.3s var(--bezier), font-size 0.3s var(--bezier);
-}
-
-.n-layout-sider--collapsed .logo {
-  padding: 8px;
-  font-size: 0;
-}
-
-.logo svg {
-  flex: 0 0 32px;
-  height: 32px;
-  margin-right: 12px;
-  transition: margin 0.3s var(--bezier);
-}
-
-.n-layout-sider--collapsed .logo svg {
-  margin-right: 0;
+  margin-top: 40px;
+  margin-bottom: 20px;
+  padding-left: 20px;
+  height: 28px;
+  color: #333;
+  font-weight: 500;
+  font-size: 20px;
 }
 </style>