+++ src/main.js
@@ -1,9 +1,7 @@
 import { createApp } from 'vue'
 import App from './App.vue'
 import router from './routes'
-import { setupNaive } from '@/plugins'
-
-import './index.css'
+import { naive } from '@/plugins'
 
 // 通用字体
 import 'vfonts/Lato.css'
@@ -13,7 +11,6 @@ import 'vfonts/FiraCode.css'
 const app = createApp(App)
 
 // 注册全局常用的 naive-ui 组件
-setupNaive(app)
-
+app.use(naive)
 app.use(router)
 app.mount('#app')