+++ src/routes/routes.js
@@ -1,0 +1,20 @@
+const mainRoutes = [{
+  name: 'home',
+  path: '/',
+  component: () => import('@/views/Home.vue'),
+}, {
+  name: 'table',
+  path: '/table',
+  component: () => import('@/views/Table.vue'),
+}]
+
+const routes = [
+  {
+    name: 'layout',
+    path: '/',
+    component: () => import("@/layout/index.vue"),
+    children: [...mainRoutes]
+  }
+]
+
+export default routes