+++ src/views/Home.vue
@@ -3,8 +3,6 @@ import HelloWorld from '@/components/HelloWorld.vue'
 </script>
 
 <template>
-  <PageFooter></PageFooter>
-  <PageHeader></PageHeader>
   <img alt="HLG logo" src="@/assets/data-talk.png" />
   <HelloWorld msg="DATA TALK" />
 </template>