+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.0 (2022-03-08)
+
+* ✨特性(Create the basic components): 创建了Header Footer Content Sidebar ([07a2e33](https://github.com/HLGhpz/Data-Talk/commit/07a2e33))
+* 📝文档(正式编写README文档): 文档更改 ([cad666b](https://github.com/HLGhpz/Data-Talk/commit/cad666b))
+* 🔧工具(Add component library): 使用naive-ui作为前端组件库 ([9107101](https://github.com/HLGhpz/Data-Talk/commit/9107101))
+* 🔧工具(Use Pnpm instead of npm;Use vite instead of vue-cli): 对程序进行了总体的重构 ([346e0dd](https://github.com/HLGhpz/Data-Talk/commit/346e0dd))
+* Initial commit ([f6c2181](https://github.com/HLGhpz/Data-Talk/commit/f6c2181))
+
+
+
 ## 0.0.0 (2022-03-05)
 
 * 🔧工具(Use Pnpm instead of npm;Use vite instead of vue-cli): 对程序进行了总体的重构 ([346e0dd](https://github.com/HLGhpz/Data-Talk/commit/346e0dd))